transformer_squared/requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.24.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
datasets>=2.14.0
huggingface-hub>=0.19.0
safetensors>=0.3.3
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
peft>=0.6.0
wandb>=0.15.0
tensorboard>=2.14.0
einops>=0.7.0
pyyaml>=6.0.0
sentencepiece>=0.1.99
pillow>=9.0.0
requests>=2.28.0 